requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
black==23.9.0
celery==5.3.4
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.1.11
django-admin-thumbnails==0.2.8
django-celery-beat==2.5.0
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-timezone-field==6.0.1
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.4.2
idna==3.4
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
